requirements.txt
adafruit_io==2.7.2
fastapi==0.111.0
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.18.2
